requirements.txt
streamlit
python-dotenv
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
langchain
langchain-google-genai
textblob
langchain_community
google-generativeai